% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{getDEE2_legacy}
\alias{getDEE2_legacy}
\title{Get DEE2 Gene Expression Data (Legacy)}
\usage{
getDEE2_legacy(
  species,
  SRRvec,
  outfile = NULL,
  metadata = NULL,
  baseURL = "http://dee2.io/cgi-bin/request.sh?",
  ...
)
}
\arguments{
\item{species}{A character string matching the species of interest.}

\item{SRRvec}{A character string or vector of SRA run accession numbers.}

\item{outfile}{An optional file name for the downloaded dataset.}

\item{metadata}{(Optional) name of R object for the meta data. Providing
the metadata will speed up performance if multiple queries are made in a
session.}

\item{baseURL}{The base URL of the service. Leave this as the default URL
unless you want to download from a 3rd party mirror.}

\item{...}{Additional parameters to be passed to download.file.}
}
\value{
a getDEE2 object.
}
\description{
This function fetches gene expression data from the DEE2 database of RNA
sequencing data and returns it as a getDEE2 object.
}
\examples{
x<-getDEE2_legacy("ecoli",c("SRR1613487","SRR1613488"))
}
\keyword{DEE2}
\keyword{RNA-seq}
\keyword{database}
