% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{getDee2Metadata}
\alias{getDee2Metadata}
\title{Get DEE2 Metadata}
\usage{
getDee2Metadata(species, outfile = NULL, ...)
}
\arguments{
\item{x}{A character string matching a species of interest.}
}
\description{
This function fetches the short metadata for the species of interest.
}
\examples{
ecoli_metadatahea<-getDee2Metadata("ecoli")
}
\keyword{metadata}
