% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{list_bundles}
\alias{list_bundles}
\title{Get a table of all completed projects at DEE2}
\usage{
list_bundles(species)
}
\arguments{
\item{species}{A character string matching a species of interest.}
}
\value{
a table of project bundles available at DEE2.io/huge
}
\description{
This function fetches a table listing all completed projects that
are available at DEE2
}
\examples{
bundles <- list_bundles("celegans")
}
\keyword{metadata}
