% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{loadFullMeta}
\alias{loadFullMeta}
\title{Load Full Metadata}
\usage{
loadFullMeta(zipname)
}
\arguments{
\item{zipname}{Path to the zipfile.}
}
\value{
a dataframe of full metadata.
}
\description{
This function loads the full metadata, which contains many fields.
}
\examples{
x<-loadQcMx("~/Downloads/Data.zip")
}
\keyword{Load}
\keyword{Metadata}
