% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{loadGeneCounts}
\alias{loadGeneCounts}
\title{Load Gene Counts}
\usage{
loadGeneCounts(zipname)
}
\arguments{
\item{zipname}{Path to the zipfile.}
}
\value{
a dataframe of gene expression counts.
}
\description{
This function loads STAR gene level counts from a downloaded zip file.
}
\examples{
x<-getDEE2("ecoli",c("SRR1613487","SRR1613488"),outfile="mydata.zip")
y<-loadGeneCounts("mydata.zip")
}
\keyword{Gene}
\keyword{Load}
