% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{loadQcMx}
\alias{loadQcMx}
\title{Load Quality Control Info}
\usage{
loadQcMx(zipname)
}
\arguments{
\item{zipname}{Path to the zipfile.}
}
\value{
a dataframe of quality control metrics.
}
\description{
This function loads quality control data. More information about the QC
metrics is available from the project github page:
https://github.com/markziemann/dee2/blob/master/qc/qc_metrics.md
}
\examples{
x<-loadQcMx("~/Downloads/Data.zip")
}
\keyword{Control}
\keyword{Load}
\keyword{QC}
\keyword{Qualiy}
