% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{loadSummaryMeta}
\alias{loadSummaryMeta}
\title{Load Summary Metadata}
\usage{
loadSummaryMeta(zipname)
}
\arguments{
\item{zipname}{Path to the zipfile.}
}
\value{
a dataframe of summary metadata.
}
\description{
This function loads the summary metadata, which are the most relevant SRA
accession numbers.
}
\examples{
x<-loadQcMx("~/Downloads/Data.zip")
}
\keyword{Load}
\keyword{Metadata}
