% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{loadTxCounts}
\alias{loadTxCounts}
\title{Load Transcript Counts}
\usage{
loadTxCounts(zipname)
}
\arguments{
\item{zipname}{Path to the zipfile.}
}
\value{
a dataframe of transcript expression counts.
}
\description{
This function loads Kallisto transcript level counts from a downloaded zip
file.
}
\examples{
x<-loadTxCounts("~/Downloads/Data.zip")
}
\keyword{Load}
\keyword{Transcript}
