% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{queryDee2}
\alias{queryDee2}
\title{Query Whether a DEE2 Dataset is Available}
\usage{
queryDee2(species, SRRvec, metadata = NULL, ...)
}
\arguments{
\item{species}{A character string matching a species of interest.}

\item{SRRvec}{A character string or vector thereof of SRA run accession
numbers.}

\item{metadata}{optional R object of DEE2 metadata to query.}

\item{...}{Additional parameters to be passed to download.file.}
}
\value{
a list of datasets that are present and absent.
}
\description{
This function sends a query to check whether a dataset is available or not.
}
\examples{
x<-queryDee2("ecoli",c("SRR1067773","SRR5350513"))
}
\keyword{query}
