% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{se}
\alias{se}
\title{Create summarizedExperiment object}
\usage{
se(x, counts = "GeneCounts")
}
\arguments{
\item{x}{a getDEE2 object.}

\item{counts}{select "GeneCounts" for STAR based gene counts, "TxCounts" for
kallisto transcript level counts or "Tx2Gene" for transcript counts
aggregated to gene level. Default is "GeneCounts"}
}
\value{
a SummarizedExperiment object
}
\description{
This function creates a SummarizedExperiment object from a getDEE2 dataset
}
\examples{
x<-getDEE2_legacy("ecoli",c("SRR1613487","SRR1613488"))
y<-se(x)
}
\keyword{SummarizedExperiment}
