% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{Tx2Gene}
\alias{Tx2Gene}
\title{Aggregate Transcript Counts to Gene-Level Counts}
\usage{
Tx2Gene(x)
}
\arguments{
\item{x}{a getDEE2 object.}
}
\description{
This function converts Kallisto transcript-level expression estimates to gene-level estimates.
}
\examples{
x<-Tx2Gene(x)
}
\keyword{Aggregate}
\keyword{gene}
\keyword{transcript}
