% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{getDEE2Metadata}
\alias{getDEE2Metadata}
\title{Get DEE2 Metadata}
\usage{
getDEE2Metadata(species, outfile = NULL, ...)
}
\arguments{
\item{species}{A character string matching a species of interest.}

\item{outfile}{Optional filename.}

\item{...}{Additional parameters to be passed to download.file.}
}
\value{
a table of metadata.
}
\description{
This function fetches the short metadata for the species of interest.
}
\examples{
ecoli_metadata <- getDEE2Metadata("ecoli")
}
\keyword{metadata}
