% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{loadGeneInfo}
\alias{loadGeneInfo}
\title{Load Gene Info}
\usage{
loadGeneInfo(zipname)
}
\arguments{
\item{zipname}{Path to the zipfile.}
}
\value{
a dataframe of gene information.
}
\description{
This function loads gene information. This information includes gene names
and lengths which is useful for downstream analysis.
}
\examples{
x<-getDEE2("ecoli",c("SRR1613487","SRR1613488"),outfile="mydata.zip")
y<-loadGeneInfo("mydata.zip")
}
\keyword{Gene}
\keyword{Load}
