% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{loadTxInfo}
\alias{loadTxInfo}
\title{Load Transcript Info}
\usage{
loadTxInfo(zipname)
}
\arguments{
\item{zipname}{Path to the zipfile.}
}
\value{
a dataframe of transcript info
}
\description{
This function loads transcript information. This information includes
transcript lengths, corresponding parent gene accession and gene symbol
that might be useful for downstream analysis.
}
\examples{
x <- getDEE2("ecoli",c("SRR1613487","SRR1613488"),outfile="mydata.zip")
y <- loadTxInfo("mydata.zip")
}
\keyword{Load}
\keyword{Transcript}
