% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{Tx2Gene}
\alias{Tx2Gene}
\title{Aggregate Transcript Counts to Gene-Level Counts}
\usage{
Tx2Gene(x)
}
\arguments{
\item{x}{a getDEE2 object.}
}
\value{
a dataframe of gene expression counts.
}
\description{
This function converts Kallisto transcript-level expression estimates to
gene-level estimates. Counts for each transcript are summed to get an
aggregated gene level score.
}
\examples{
x<-getDEE2("celegans",c("SRR3657716","SRR3657717"))
x<-Tx2Gene(x)
head(x$Tx2Gene)
}
\keyword{Aggregate}
\keyword{gene}
\keyword{transcript}
